+++ ci/run_applications.sh
@@ -6,7 +6,7 @@ ls
 PYTHON=python3
 PYTHON_PREPROC="export MPLBACKEND=agg; export hIPPYlibDeprecationWarning=error; cd /home/fenics/hippylib; $PYTHON -c 'import hippylib'; cd -;"
 
-IMAGE=hippylib/fenics:latest
+
 
 DOCKER="docker run --rm -v $(pwd):/home/fenics/hippylib -w"