+++ ci/run_unittest.sh
@@ -5,8 +5,6 @@ set -ev
 
 PYTHON=python3
 
-IMAGE=hippylib/fenics:latest
-
 DOCKER="docker run --rm -v $(pwd):/home/fenics/hippylib -w"
 
 ${DOCKER} /home/fenics/hippylib/hippylib/test $IMAGE "$PYTHON -m unittest discover -v"