+++ hippylib/algorithms/linSolvers.py
@@ -1,0 +1,38 @@
+# Copyright (c) 2016-2018, The University of Texas at Austin 
+# & University of California--Merced.
+# Copyright (c) 2019, The University of Texas at Austin 
+# University of California--Merced, Washington University in St. Louis.
+#
+# All Rights reserved.
+# See file COPYRIGHT for details.
+#
+# This file is part of the hIPPYlib library. For more information and source code
+# availability see https://hippylib.github.io.
+#
+# hIPPYlib is free software; you can redistribute it and/or modify it under the
+# terms of the GNU General Public License (as published by the Free
+# Software Foundation) version 2.0 dated June 1991.
+
+import dolfin as dl
+
+
+def PETScKrylovSolver(comm, method, preconditioner):
+    try:
+        out = dl.PETScKrylovSolver(comm, method, preconditioner)
+    except:
+        out = dl.PETScKrylovSolver(method, preconditioner)
+        
+    return out
+
+def _PETScLUSolver_set_operator(self, A):
+    if hasattr(A, 'mat'):
+        self.ksp().setOperators(A.mat())
+    else:
+        self.ksp().setOperators(dl.as_backend_type(A).mat())
+    
+
+def PETScLUSolver(comm, method='default'):
+    out = dl.PETScLUSolver(comm, method)
+    
+    if not hasattr(out, 'set_operator'):
+        out.set_operator = _PETScLUSolver_set_operator